pyproject.toml
[tool.poetry]
name = "reflexion-writer"
version = "0.1.0"
description = ""
authors = ["Caio Bonfante <caio.bonfante@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
langgraph = "^0.0.38"
python-dotenv = "^1.0.1"
black = "^24.4.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
